requirements.txt
Django~=4.1.13
django-environ
djangorestframework~=3.14.0
djongo==1.3.6
django-cors-headers
google-generativeai
pillow
pymongo==3.12.1
pytz
python-dotenv~=1.0.0
vertexai

pypdf2~=3.0.1